move ConsumptionVerifier
