+++ tests/Parallafka.Tests/Commits/CommitTestsBase.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Parallafka.KafkaConsumer;
+using Parallafka.Tests.Helpers;
 using Parallafka.Tests.OrderGuarantee;
 using Xunit;
 using Xunit.Abstractions;
