+++ tests/Parallafka.Tests/OrderGuarantee/OrderGuaranteeTestBase.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Parallafka.KafkaConsumer;
+using Parallafka.Tests.Helpers;
 using Xunit;
 using Xunit.Abstractions;
