+++ tests/Parallafka.Tests/Performance/ThroughputTestsBase.cs
@@ -4,6 +4,7 @@ using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
 using Parallafka.KafkaConsumer;
+using Parallafka.Tests.Helpers;
 using Parallafka.Tests.OrderGuarantee;
 using Xunit;
 using Xunit.Abstractions;
